requirements.txt
certifi==2019.6.16
cloudinary==1.17.0
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.7
isort==4.3.4
mccabe==0.6.1
mock==3.0.5
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
